WIP - Some of the 3D stuff in place
I have the code in place now that allows me to choose three different
ways to control the robot, including the Backhoe mode as well as a 3D
move the arm.

The 3D stuff is just barely limping along and causes major arm movments.
Need to solve.  Partially need to verify that the desired move is doable
and stop it before it tries...