VB Debug app added
Decided to also include my VB application I use to forward messages from
Commander to Robot.  The way I used this is to move the XBee from the
robot to USB one connected to PC, Connected the FTDI cable to arbotix
board.  Then in this VB app selected the comm ports and did a connect...